Move email handler/authentication to common loc
